Normalised sections in files
This normalisation makes it easy to actually use the include files in any InnoSetup script without worrying about the order in which to include files and whether an included files leaves a [Code] section open. Leaked [Code] sections change the comment syntax for everything that follows. This can lead to a big mess if not handled carefully. My change never leaves a [Code] section open at the end of a file which avoids this kind of problem completely.